+++ src/main/java/net/ktccenter/campusApi/dto/lite/administration/LiteParametreInstitutionDTO.java
@@ -14,7 +14,7 @@ public class LiteParametreInstitutionDTO {
     private Long id;
     private String bareme;
     private String devise;
-    private Integer dureeCours;
+    private double dureeCours;
 
     public LiteParametreInstitutionDTO(ParametreInstitution parametres) {
         this.id = parametres.getId();
